+++ CHANGELOG.md
@@ -20,6 +20,8 @@ This release marked the change of PHPWord license from LGPL 2.1 to LGPL 3; new r
 - RTF: Ability to add links and page breaks in RTF - @ivanlanin GH-196
 - ListItemRun: Remove fontStyle parameter because ListItemRun is inherited from TextRun and TextRun doesn't have fontStyle - @ivanlanin
 - Config: Ability to use a config file to store various common settings - @ivanlanin GH-200
+- ODT: Enable inline font style in TextRun - @ivanlanin
+- ODT: Enable underline, strike/doublestrike, smallcaps/allcaps, superscript/subscript font style - @ivanlanin
 
 ### Bugfixes